refactor: extract call flow logic
